requirements.txt
# NIDS Requirements
# Core dependencies
scapy>=2.5.0
yara-python>=1.8.0
aiohttp>=3.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
aiosqlite>=0.19.0

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0